requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
httpx==0.25.2
apscheduler==3.10.4
python-dotenv==1.0.0
